feat(FEV-886): use drag&snap manager as a HOC
