+++ src/components/drag-and-snap-manager/drag-and-snap-manager.scss
@@ -1,0 +1,3 @@
+.draggableContainer {
+  position: absolute;
+}
