+++ src/components/pip/index.ts
@@ -1,1 +1,2 @@
-export * from "./pip";
+export * from "./pip-child";
+export * from "./pip-parent";
