--- src/components/pip/pip.tsx
@@ -1,175 +1,0 @@
-import {h, createRef, Component} from 'preact';
-import * as styles from './pip.scss';
-import {Position, Animaitons} from '../../enums';
-import {icons} from '../../icons';
-import {DragAndSnapManager} from '../../dragAndSnapManager';
-import {GuiClientRect} from '../../types';
-const {connect} = KalturaPlayer.ui.redux;
-const {utils, reducers} = KalturaPlayer.ui;
-const {Icon} = KalturaPlayer.ui.components;
-
-const INITIAL_POSITION = -150;
-
-const mapStateToProps = (state: Record<string, any>) => ({
-  playerHover: state.shell.playerHover,
-  guiClientRect: state.shell.guiClientRect,
-  prePlayback: state.engine.prePlayback
-});
-
-interface PIPComponentOwnProps {
-  childPlayer: KalturaPlayerTypes.Player;
-  childSizePercentage: number;
-  inverse?: boolean;
-  position: Position;
-  hide: () => void;
-  onSideBySideSwitch?: () => void;
-  onInversePIP?: () => void;
-  dragAndSnapManager?: DragAndSnapManager;
-  animation: Animaitons;
-}
-interface PIPComponentConnectProps {
-  guiClientRect?: GuiClientRect;
-  playerHover?: boolean;
-  prePlayback?: boolean;
-}
-interface PIPComponentState {
-  isDragging: boolean;
-}
-type PIPComponentProps = PIPComponentOwnProps & PIPComponentConnectProps;
-@connect(mapStateToProps, utils.bindActions(reducers.shell.actions))
-export class Pip extends Component<PIPComponentProps, PIPComponentState> {
-  videoContainerRef = createRef<HTMLDivElement>();
-  pipContainerRef = createRef<HTMLDivElement>();
-  constructor(props: PIPComponentProps) {
-    super(props);
-    this.state = {
-      isDragging: false
-    };
-  }
-
-  componentDidMount() {
-    const {childPlayer, dragAndSnapManager} = this.props;
-    this.videoContainerRef?.current?.appendChild(childPlayer.getView());
-    if (dragAndSnapManager && this.pipContainerRef.current) {
-      dragAndSnapManager.setDraggableContainer(this.pipContainerRef.current);
-      dragAndSnapManager.onStartDrag(this._onStartDrag);
-      dragAndSnapManager.onStopDrag(this._onStopDrag);
-    }
-  }
-
-  componentDidUpdate() {
-    const {dragAndSnapManager, guiClientRect} = this.props;
-    if (dragAndSnapManager && guiClientRect) {
-      dragAndSnapManager.setGuiClientRect(guiClientRect);
-    }
-  }
-
-  private _handleClick = (fn: Function) => (e: Event) => {
-    e.stopPropagation();
-    this.props.dragAndSnapManager?.destroy();
-    fn();
-  };
-
-  private _renderHoverButton() {
-    const {onSideBySideSwitch, hide, onInversePIP, inverse, playerHover} = this.props;
-    if (!playerHover || inverse || this.state.isDragging) {
-      return null;
-    }
-    return (
-      <div>
-        <div className={styles.innerButtons}>
-          <div className={styles.iconContainer} onMouseUp={this._handleClick(onInversePIP!)}>
-            <Icon id="dualscreen-pip-swap" height={24} width={24} path={icons.SWAP_ICON_PATH} />
-          </div>
-          <div className={styles.iconContainer} onMouseUp={this._handleClick(onSideBySideSwitch!)}>
-            <Icon id="dualscreen-pip-side-by-side" height={24} width={24} path={icons.SIDE_BY_SIDE_ICON_PATH} />
-          </div>
-        </div>
-        <div onMouseUp={this._handleClick(hide)} role="button" className={styles.hideContainer}>
-          <div className={styles.iconContainer}>
-            <Icon id="dualscreen-pip-hide" height={16} width={16} path={icons.HIDE_ICON_PATH} />
-          </div>
-          Hide
-        </div>
-      </div>
-    );
-  }
-
-  private _onStartDrag = () => {
-    this.setState({
-      isDragging: true
-    });
-  };
-
-  private _onStopDrag = () => {
-    this.setState({
-      isDragging: false
-    });
-  };
-
-  render(props: PIPComponentProps) {
-    const styleClass = [];
-
-    if (!props.inverse) {
-      styleClass.push(styles.childPlayer);
-
-      if (props.playerHover) {
-        styleClass.push(styles.playerHover);
-      }
-
-      if (this.state.isDragging) {
-        styleClass.push(styles.dragging);
-      }
-    }
-
-    if (!props.prePlayback && props.animation) {
-      switch (props.animation) {
-        case Animaitons.Fade:
-          styleClass.push(styles.animatedFade);
-          break;
-        case Animaitons.ScaleRight:
-          styleClass.push(styles.animatedScaleRight);
-          break;
-        case Animaitons.ScaleLeft:
-          styleClass.push(styles.animatedScaleLeft);
-          break;
-      }
-    }
-
-    const pipContainerStyles: Record<string, number> = {};
-
-    if (!props.inverse) {
-      switch (props.position) {
-        case Position.BottomRight:
-          pipContainerStyles.bottom = props.prePlayback ? INITIAL_POSITION : 0;
-          pipContainerStyles.right = 0;
-          break;
-        case Position.BottomLeft:
-          pipContainerStyles.bottom = props.prePlayback ? INITIAL_POSITION : 0;
-          pipContainerStyles.left = 0;
-          break;
-        case Position.TopRight:
-          pipContainerStyles.top = props.prePlayback ? INITIAL_POSITION : 0;
-          pipContainerStyles.right = 0;
-          break;
-        case Position.TopLeft:
-          pipContainerStyles.top = props.prePlayback ? INITIAL_POSITION : 0;
-          pipContainerStyles.left = 0;
-      }
-    }
-
-    const height: number = props.inverse ? props.guiClientRect!.height : (props.guiClientRect!.height * props.childSizePercentage) / 100;
-    const width: number = props.inverse ? props.guiClientRect!.width : (height * 16) / 9;
-    const videoContainerStyles = {
-      height: `${height + 'px'}`,
-      width: `${width + 'px'}`
-    };
-
-    return (
-      <div className={styleClass.join(' ')} ref={this.pipContainerRef} style={pipContainerStyles}>
-        <div className={styles.videoContainer} style={videoContainerStyles} ref={this.videoContainerRef} />
-        {this._renderHoverButton()}
-      </div>
-    );
-  }
-}
